packages/pykit-logging/pyproject.toml
[project]
name = "pykit-logging"
description = "Structured logging with structlog integration"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "structlog",
]

[project.optional-dependencies]
otlp = [
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29.0",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pykit_logging"]
